refactor convert_mistral_to_gguf.py in convert_hf_to_gguf.py
