+++ gguf-py/gguf/utility.py
@@ -120,7 +120,7 @@ class SafetensorRemote:
         """
         # case 1: model has only one single model.safetensor file
         is_single_file = cls.check_file_exist(f"{cls.BASE_DOMAIN}/{model_id}/resolve/main/model.safetensors")
-        is_single_file_consolidated = cls.check_file_exist(f"{cls.BASE_DOMAIN}/{model_id}/resolve/main/consolidated.safetensors", user_agent="convert_mistral_to_gguf")
+        is_single_file_consolidated = cls.check_file_exist(f"{cls.BASE_DOMAIN}/{model_id}/resolve/main/consolidated.safetensors")
         if is_single_file:
             url = f"{cls.BASE_DOMAIN}/{model_id}/resolve/main/model.safetensors"
             return cls.get_list_tensors(url)
@@ -242,7 +242,7 @@ class SafetensorRemote:
         return response.content[slice(size if size > -1 else None)]
 
     @classmethod
-    def check_file_exist(cls, url: str, user_agent="convert_hf_to_gguf") -> bool:
+    def check_file_exist(cls, url: str) -> bool:
         """
         Check if a file exists at the given URL.
         Returns True if the file exists, False otherwise.
@@ -255,7 +255,7 @@ class SafetensorRemote:
             raise ValueError(f"Invalid URL: {url}")
 
         try:
-            headers = cls._get_request_headers(user_agent=user_agent)
+            headers = cls._get_request_headers()
             headers["Range"] = "bytes=0-0"
             response = requests.head(url, allow_redirects=True, headers=headers)
             # Success (2xx) or redirect (3xx)
@@ -264,9 +264,9 @@ class SafetensorRemote:
             return False
 
     @classmethod
-    def _get_request_headers(cls, user_agent="convert_hf_to_gguf") -> dict[str, str]:
+    def _get_request_headers(cls) -> dict[str, str]:
         """Prepare common headers for requests."""
-        headers = {"User-Agent": user_agent}
+        headers = {"User-Agent": "convert_hf_to_gguf"}
         if os.environ.get("HF_TOKEN"):
             headers["Authorization"] = f"Bearer {os.environ['HF_TOKEN']}"
         return headers
