+++ requirements.txt
@@ -10,5 +10,4 @@
 -r ./requirements/requirements-convert_hf_to_gguf_update.txt
 -r ./requirements/requirements-convert_llama_ggml_to_gguf.txt
 -r ./requirements/requirements-convert_lora_to_gguf.txt
--r ./requirements/requirements-convert_mistral_to_gguf.txt
 -r ./requirements/requirements-tool_bench.txt
