+++ requirements/requirements-all.txt
@@ -11,7 +11,6 @@
 -r ./requirements-convert_hf_to_gguf_update.txt
 -r ./requirements-convert_legacy_llama.txt
 -r ./requirements-convert_llama_ggml_to_gguf.txt
--r ./requirements-convert_mistral_to_gguf.txt
 -r ./requirements-tool_bench.txt
 
 -r ./requirements-gguf_editor_gui.txt
