+++ requirements/requirements-convert_hf_to_gguf.txt
@@ -1,6 +1,7 @@
 mistral-common>=1.8.3
 
 -r ./requirements-convert_legacy_llama.txt
+mistral-common>=1.8.3
 --extra-index-url https://download.pytorch.org/whl/cpu
 torch~=2.2.1; platform_machine != "s390x"
