--- requirements/requirements-convert_mistral_to_gguf.txt
@@ -1,13 +1,0 @@
-numpy<2.0.0
-gguf>=0.1.0
-protobuf>=4.21.0,<5.0.0
-mistral-common>=1.8.0
-safetensors>=0.5.3
-huggingface_hub>=0.23.2
-
---extra-index-url https://download.pytorch.org/whl/cpu
-torch~=2.2.1; platform_machine != "s390x"
-
-# torch s390x packages can only be found from nightly builds
---extra-index-url https://download.pytorch.org/whl/nightly
-torch>=0.0.0.dev0; platform_machine == "s390x"
